requirements.txt
Django==4.2.13
django-environ==0.11.2
psycopg2-binary==2.9.9
pgvector==0.2.5
djangorestframework==3.15.1
drf-nested-routers==0.94.1
drf-yasg==1.21.7
django-filter==24.2
python-telegram-bot==21.1.1
django-pglocks==1.0.4
django-mptt==0.16.0
markdown2==2.5.1
beautifulsoup4==4.12.3
openai==1.28.1
groq==0.6.0
ollama==0.4.4
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
numpy==1.26.4
aiohttp==3.9.5
navec==0.10.0
langid==1.1.6
celery==5.4.0
celery-pool-asyncio==0.2.0
amqp==5.2.0
propcache==0.2.1
httpcore==1.0.6
redis==5.0.7
lxml==4.9.3
pillow==11.0.0
requests==2.32.3

pytest==8.2.2
pytest-mock==3.14.0
pytest-asyncio==0.23.7
pytest-django==4.8.0
factory_boy==3.3.1